Segregate more of the printing methods
