+++ link-grammar/dict-file/read-dict.h
@@ -23,6 +23,7 @@ bool read_dictionary(Dictionary dict);
 
 Dict_node * lookup_list(const Dictionary dict, const char *s);
 bool boolean_lookup(Dictionary dict, const char *s);
+Dict_node * dictionary_lookup_wild(Dictionary dict, const char *s);
 void free_lookup(Dict_node *llist);
 void free_insert_list(Dict_node *ilist);
 void insert_list(Dictionary dict, Dict_node * p, int l);
